backend/requirements.txt.txt
fastapi
uvicorn
pydantic
pdfplumber
python-docx
spacy